feat: extract common ticketing components to base (#98016)
