+++ ui/components/task-components/swipe/swipe.tsx
